now able to read 30GB data on each process

+++ src/c/dynamic_2d_array.h
@@ -5,7 +5,7 @@
 #include <stdlib.h>
              
 //typedef double integer;
-typedef int integer;
+typedef unsigned long long integer;
 //typedef short sint;
 typedef unsigned short sint;
 typedef char cint;
@@ -14,11 +14,11 @@ typedef float real;
 //typedef double real;
 #define MPI_REALNUM MPI_FLOAT
 
-integer **allocate_dynamic_2d_array_integer(int nrows, int ncols);
+integer **allocate_dynamic_2d_array_integer(integer nrows, integer ncols);
 void free_dynamic_2d_array_integer(integer** array_dynamic);
 void print_matrix_integer(integer** array_dynamic, int nrows, int ncols, char* fmt_string);
 
-sint **allocate_dynamic_2d_array_sint(int nrows, int ncols);
+sint **allocate_dynamic_2d_array_sint(integer nrows, integer ncols);
 void free_dynamic_2d_array_sint(sint** array_dynamic);
 void print_matrix_sint(sint** array_dynamic, int nrows, int ncols, char* fmt_string);
 
@@ -30,9 +30,9 @@ real **allocate_dynamic_2d_array_real(int nrows, int ncols);
 void free_dynamic_2d_array_real(real** array_dynamic);
 void print_matrix_real(real** array_dynamic, int nrows, int ncols, char* fmt_string);
 
-integer** allocate_dynamic_2d_array_integer(int nrows, int ncols) {
+integer** allocate_dynamic_2d_array_integer(integer nrows, integer ncols) {
     /* here is the method to correct the non-contiguous memory problem */
-    int i;
+    integer i;
     integer** array_dynamic=(integer**)malloc(nrows*sizeof(integer*));
     integer* data=(integer*)malloc(nrows*ncols*sizeof(integer));
     for (i=0; i<nrows; i++){
@@ -57,9 +57,9 @@ void print_matrix_integer(integer** array_dynamic, int nrows, int ncols, char* f
     }
 }
 
-sint** allocate_dynamic_2d_array_sint(int nrows, int ncols) {
+sint** allocate_dynamic_2d_array_sint(integer nrows, integer ncols) {
     /* here is the method to correct the non-contiguous memory problem */
-    int i;
+    integer i;
     sint** array_dynamic=(sint**)malloc(nrows*sizeof(sint*));
     sint* data=(sint*)malloc(nrows*ncols*sizeof(sint));
     for (i=0; i<nrows; i++){
